Organizational Changes
This will make it easier to implement the Phoenix framework later on
